+++ CHANGELOG.md
@@ -2,13 +2,18 @@
 
 ## [Unreleased]
 
+### Added
+
+- Added `packet_header_size()` for applications that need to know the size of internal type
+  `pcap_pkthdr` (for instance to calculate exact send queue sizes).
+
+want to precalculate exact queue sizes.
+
 ## [2.2.0] - 2024-09-01
 
 ### Added
 
 - Added an implementation of `AsFd` on `Capture<Active>` on non-Windows.
-- Added `SendQueue::packet_header_size()` for applications that want to
-  precalculate exact queue sizes.
 
 ## [2.1.0] - 2024-08-27